move some functions into util.py
